changed passwd to long int

+++ client-udp.c
@@ -9,7 +9,7 @@
 
 int main(int argc, char const *argv[]) {
     bool hasPasswd = false;
-    int passwd = 1;
+    long int passwd = 1;
     if(argc < 3){
         fprintf(stderr, "Two few arguments given\n");
         exit(EXIT_FAILURE);
